Module Online Modifications Added (Bugged, Read Desc)
This module is bugged so I commented the working code out, everything else works well.
It is supposed to work by taking all unknown objects and adding a parent to them, but creates too many parents in the hierarchy. Working on fix soon.
Another thing to add to this module is the ability to add many more relationships to all the words, working on that later.