Added Solve Ambiguity Button (READ DESC).
This is a WIP method that tries to disambiguate words that are descendants of Object using ChatGPT. It takes a word like "bow" and outputs bow1 is-a gesture, bow2 is-a weapon, etc. Next I will try to disambiguate from files.
NOTE: This takes a lot of tokens to run, so be careful even on medium sized datasets.